refactor alchemy methods into its own file under chain-data-utils
cleaner/more consistent this way
